Do not offer pushing to buildout.coredev is PLONE_RELEASER_MULTI_PACKAGES env var is set.
When doing releases of several packages in a row, it is better to not push, because this means a needlessly high load on Jenkins.
Also, if you create a release and immediately push, then Jenkins will not find the new release yet and fail:
it takes a few minutes for the release to be propagated to all PyPI mirrors.
Pushing still seems the best default, but let's have an easy way to not push.
We look for environment variable PLONE_RELEASER_MULTI_PACKAGES.
Best is value 0 or 1, but we try not to break if you have something else.