Add VersionsFile.rewrite.
This takes the parsed 'extends' and 'versions' and writes them back.
This normalizes some stuff, so it will not be exactly the same.

In the end we want to take versions.cfg and translate it to constraints.txt.
But translating versions.cfg to its own should work as well, and is a good first step.