manage changelog: add --package option.
I see a problem getting the changelog from plone.restapi, so it is useful to be able to restrict the command to this package.
Found the problem: the start version of plone.restapi is not included in its CHANGES.rst on the main branch.
Explained this corner case in the new help text.